+++ src/Tests/TestApp1/EvolutionAlgorithmFactory.cs
@@ -12,55 +12,55 @@ namespace TestApp1;
 
 public static class EvolutionAlgorithmFactory
 {
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_Xor()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_Xor()
     {
         var experimentFactory = new XorExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_Binary6()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_Binary6()
     {
         var experimentFactory = new BinarySixMultiplexerExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_Binary11()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_Binary11()
     {
         var experimentFactory = new BinaryElevenMultiplexerExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_GenerativeSinewave()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_GenerativeSinewave()
     {
         var experimentFactory = new GenerativeFnRegressionExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_CartSinglePole()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_CartSinglePole()
     {
         var experimentFactory = new CartSinglePoleExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_CartDoublePole()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_CartDoublePole()
     {
         var experimentFactory = new CartDoublePoleExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    public static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm_PreyCapture()
+    public static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm_PreyCapture()
     {
         var experimentFactory = new PreyCaptureExperimentFactory();
         return CreateNeatEvolutionAlgorithm(experimentFactory);
     }
 
-    private static NeatEvolutionAlgorithm<double> CreateNeatEvolutionAlgorithm(
+    private static NeatEvolutionAlgorithm<float> CreateNeatEvolutionAlgorithm(
         INeatExperimentFactory experimentFactory)
     {
         string jsonConfigFilename = $"experiments-config/{experimentFactory.Id}.config.json";
 
         // Create an instance of INeatExperiment, configured using the supplied json config.
-        INeatExperiment<double> neatExperiment = experimentFactory.CreateExperiment<double>(jsonConfigFilename);
+        INeatExperiment<float> neatExperiment = experimentFactory.CreateExperiment<float>(jsonConfigFilename);
 
         // Create a NeatEvolutionAlgorithm instance ready to run the experiment.
         var ea = NeatEvolutionAlgorithmFactory.CreateEvolutionAlgorithm(neatExperiment);
